+++ src/RayTracer/Object/Camera/Camera.hpp
@@ -14,12 +14,12 @@
 namespace RayTracer {
     class Camera: public Object, public ICamera {
         public:
-            Camera(int width, int height, double fov);
+            Camera(int width, int height, float fov);
             virtual ~Camera() = default;
             Maths::Rect3D getScreen() const { return _screen; };
             void setScreen(Maths::Rect3D screen) { _screen = screen; };
 
-            Maths::Ray ray(double x, double y) const override;
+            Maths::Ray ray(float x, float y) const override;
         private:
             Maths::Rect3D _screen;
     };
